feat: Added create user mutation
